Variation and Game are now the same object
