+++ src/game/action/game.rs
@@ -1,5 +1,3 @@
-use crate::game::{compiles_to::CompilesTo, ChessAction, Game};
-
 pub enum GameAction {
     // Seeks forward/back by specified number of actions
     Seek(usize),
