--- src/game/variation_builder.rs
@@ -1,5 +1,0 @@
-
-
-pub struct VariationBuilder {
-
-}
